+++ nes_emu/src/nes_emu/ppu.cpp
@@ -191,6 +191,12 @@ NES_Address PPU::sprite_pattern_address(NES_Byte tile, int y_offset) const {
     return address;
 }
 
+void PPU::observe_dummy_sprite_fetch(PictureBus& bus) const {
+    const NES_Address address = sprite_pattern_address(0xff, 0);
+    bus.read(address);
+    bus.read(address + 8);
+}
+
 void PPU::prefetch_scanline_sprite_rows(PictureBus& bus) {
     std::fill(
         scanline_sprite_rows.begin(),
@@ -534,6 +540,13 @@ void PPU::cycle(PictureBus& bus) {
                 data_address &= ~0x41f;
                 data_address |= temp_address & 0x41f;
                 invalidate_background_tile_cache();
+            }
+            else if (
+                cycles == SCANLINE_VISIBLE_DOTS + 4 &&
+                is_showing_sprites &&
+                bus.requires_sprite_fetch_address_observations()
+            ) {
+                observe_dummy_sprite_fetch(bus);
             }
 
 //                 if (cycles > 257 && cycles < 320)